agent/requirements.txt
psutil
requests
python-dotenv
wmi
